Reorganised /tests/ dir
